teste/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
packages = [{include = "backend"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref (==3.9.2)",
    "django (==5.2.6)",
    "pillow (==11.3.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "sqlparse (==0.5.3)",
    "django-cpf (>=0.1.0,<0.2.0)",
    "django-phonenumber-field[phonenumbers] (>=8.1.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.13.1,<0.14.0)"
]
